#2 Moved responsibility to separate component
